Wired up torch modules, added option to test correctness directly against E3NN.
